BingoSim.Board

Description

This module contains the data types we'll be using, which is really just a way to represent and operate on boards.

For our purposes, we only need random boards, so it's sufficient to identify a board position by its index and then represent a board by a bit vector. Compared to a list or some other kind of sequence, this lets us use less space and perform operations on boards in fewer instructions.

Since we're using a Word64 but only representing boards with 36 grid spaces, we only use the 36 least significant bits: 0xf_ffff_ffff

We represent boards in row-major order, so each multiple of 6 bits is a row.

Boards

newtype Board

The type of a 6x6 game board, as a bit sequence. Only the 36 least significant bits are meaningful.

Constructors

Board Word64

Bingos

The main operation we want to do on a board is check whether it has a bingo or not.

hasBingo :: Board -> Maybe BingoLocation

Check whether there's a bingo, and if so, return the BingoLocation corresponding to where the bingo occured (i.e., which row, column, or diagonal).

In the case where there are multiple bingos, returns one of them arbitrarily.

>>> hasBingo (Board 0x123456789)
Nothing
>>> hasBingo (Board 0x3f)
Just Row6

data BingoLocation

An enum of where a bingo could occur on a 6x6 board.

Constructors

Dia1
Top left to bottom right

Dia2
Top right to bottom left

Row1
Top row

Row2

Row3

Row4

Row5

Row6
Bottom row

Col1
Far left column

Col2

Col3

Col4

Col5

Col6
Far right column

Board Constants

These are a handful of bit masks to project out the rows, columns, and diagonals of a board.

dia1 :: Word64

Diagonal from top left to bottom right

dia2 :: Word64

Diagonal from top right to bottom left

row1 :: Word64

Top row

row2 :: Word64

Second row from top

row3 :: Word64

Third row from top

row4 :: Word64

Fourth row from top

row5 :: Word64

Fifth row from top

row6 :: Word64

Bottom row

col1 :: Word64

Far left column

col2 :: Word64

Second column from left

col3 :: Word64

Third column from left

col4 :: Word64

Fourth column from left

col5 :: Word64

Fifth column from left

col6 :: Word64

Far right column